Move ContextMenu logic to PageCell component.
